multi face detect

+++ src/main/java/org/alking/tecent/ai/face/MultiFaceReply.java
@@ -9,43 +9,43 @@ public class MultiFaceReply extends BaseReply<MultiFaceReply.Data> {
 
     public static class FaceItem {
 
-        private int x1;
+        private Double x1;
 
-        private int y1;
+        private Double y1;
 
-        private int x2;
+        private Double x2;
 
-        private int y2;
+        private Double y2;
 
-        public int getX1() {
+        public Double getX1() {
             return x1;
         }
 
-        public void setX1(int x1) {
+        public void setX1(Double x1) {
             this.x1 = x1;
         }
 
-        public int getY1() {
+        public Double getY1() {
             return y1;
         }
 
-        public void setY1(int y1) {
+        public void setY1(Double y1) {
             this.y1 = y1;
         }
 
-        public int getX2() {
+        public Double getX2() {
             return x2;
         }
 
-        public void setX2(int x2) {
+        public void setX2(Double x2) {
             this.x2 = x2;
         }
 
-        public int getY2() {
+        public Double getY2() {
             return y2;
         }
 
-        public void setY2(int y2) {
+        public void setY2(Double y2) {
             this.y2 = y2;
         }
